ref: restructured files for relative css modules
