+++ src/pages/auth-page/CreateProfile.js
@@ -1,8 +1,9 @@
 import React from "react";
+import authpagesStyle from "./authpages.module.scss";
 
 export const CreateProfile = () => {
 	return (
-		<main>
+		<main className={authpagesStyle.page}>
 			<h1>What Should We Call You?</h1>
 		</main>
 	);
