+++ src/pages/auth-page/authpages.module.scss
@@ -1,4 +1,4 @@
-.signPage{
+.page{
     background-color: #272727;
     min-height: 100vh;
     display: flex;
